added types to the root stuff
