+++ src/index.tsx
@@ -5,11 +5,11 @@ import store from './store'
 
 import { library } from '@fortawesome/fontawesome-svg-core'
 import { faSearch } from '@fortawesome/free-solid-svg-icons'
-import StockTracker from './stockTracker'
-import { bootstrap } from './bootstrap'
 import 'ress'
-import './index.css'
+import { bootstrap } from './bootstrap'
 import './colours.css'
+import './index.css'
+import StockTracker from './stockTracker'
 
 library.add(faSearch)
