+++ src/store.ts
@@ -1,7 +1,7 @@
-import { createStore, applyMiddleware } from 'redux'
+import { applyMiddleware, createStore } from 'redux'
 import thunkMiddleware from 'redux-thunk'
-import rootReducer from './reducer'
 import api from './api'
+import rootReducer from './reducer'
 
 const store = createStore(rootReducer, applyMiddleware(thunkMiddleware.withExtraArgument(api)))
